chore(commands): refactor command handler, give UI options to clear command cooldowns (#1947)
